requirements-dev.txt
numpy==2.0.0
pandas
psycopg2
sqlalchemy
requests
fastapi
uvicorn
PyJWT